Closing in on a solution
Moved the contents of test/main.py to test/test_DataFrameModelManager.py
that way I can actually run the tests after a fresh install. Commented
out some requirements in setup.py.
