+++ setup.py
@@ -7,26 +7,30 @@ import os
 import re
 import sys
 
-has_qt4 = True
-try:
-    # sip is only needed for PyQt4, they should be imported together.
-    # If we can let's remove all of the references to PyQt or Pyside in favor
-    # of qtpy.
-    import PyQt4
-    import sip
-except ImportError as e:
-    has_qt4 = False
+# TODO: Remove the commented out loop below. Users should take care of the
+# PySide or PyQt requirements before they install that way they can decide what
+# works best for them. Then we can just use qtpy through the compat.py file
+# to handle whatever Qt solution they picked.
 
-try:
-    import PySide
-except ImportError as e:
-    # TODO: try to import PySide.
-    if not has_qt4:
-        # We know we failed to import PyQt4/sip...
-        # And we failed to import pyside.
-        raise ImportError("\n\ninstall PyQt4 and sip or PySide")
-    else:
-        print("Using PyQt4")
+# has_qt4 = True
+# try:
+#     # sip is only needed for PyQt4, they should be imported together.
+#     # If we can let's remove all of the references to PyQt or Pyside in favor
+#     # of qtpy.
+#     import PyQt4
+#     # import sip
+# except ImportError as e:
+#     has_qt4 = False
+#
+# try:
+#     import PySide
+# except ImportError as e:
+#     if not has_qt4:
+#         # We know we failed to import PyQt4/sip...
+#         # And we failed to import pyside.
+#         raise ImportError("\n\nPlease install PyQt4 and sip or PySide")
+#     else:
+#         print("Using PyQt4")
 
 here = os.path.abspath(os.path.dirname(__file__))
 
@@ -53,7 +57,7 @@ library for Python) with Qt."""
 try:
     long_description = read('README.md')
 except:
-    long_description = "See README.md where installed"
+    long_description = "See README.md where installed."
 
 
 class PyTest(TestCommand):
@@ -68,8 +72,9 @@ class PyTest(TestCommand):
         sys.exit(errcode)
 
 
-tests_require = ["pandas == 0.17.1", 'easygui',  'pyside',
+tests_require = ["pandas == 0.17.1", 'easygui', 'pyqt',#'pyside',
                  'pytest', 'pytest-cov', 'pytest-qt']# 'python-magic==0.4.6']
+
 setup(
     name='qtpandas',
     version=__version__,
